(PC-30978)[PRO] feat: Create a new row for collective bookable offers in offers table.
