+++ pro/src/pages/Offers/OffersTable/CollectiveOffersTable/CollectiveOffersTableHead/CollectiveOffersTableHead.module.scss
@@ -1,0 +1,10 @@
+@use "styles/mixins/_a11y.scss" as a11y;
+
+.visually-hidden {
+  @include a11y.visually-hidden;
+}
+
+.expiration-date-th,
+.offer-reference-th {
+  padding: 0;
+}
